restructure domain core + code issues fixed
