+++ components/ui/button.tsx
@@ -1,6 +1,6 @@
 import { forwardRef } from 'react';
 import Link from 'next/link';
-import { cn } from '../../lib/client/utils';
+import { cn } from '../../domain_core/infrastructure/utils';
 
 export interface ButtonProps
   extends React.ButtonHTMLAttributes<HTMLButtonElement> {
