+++ components/ui/label.tsx
@@ -3,7 +3,7 @@
 import * as React from 'react';
 import * as LabelPrimitive from '@radix-ui/react-label';
 
-import { cn } from '@/lib/client/utils';
+import { cn } from '@/domain_core/infrastructure/utils';
 
 function Label({
   className,
