+++ components/ui/radio-group.tsx
@@ -4,7 +4,7 @@ import * as React from 'react';
 import * as RadioGroupPrimitive from '@radix-ui/react-radio-group';
 import { CircleIcon } from 'lucide-react';
 
-import { cn } from '@/lib/client/utils';
+import { cn } from '@/domain_core/infrastructure/utils';
 
 function RadioGroup({
   className,
