+++ domain_core/infrastructure/utils.ts
