+++ domain_core/model/domain_model.ts
@@ -19,6 +19,7 @@ export interface CreateUserInput {
 
 export interface UpdateUserInput {
   name?: string;
+  email?: string;
   role?: UserRole;
 }
