+++ instrumentation.ts
@@ -1,7 +1,9 @@
 /* eslint-disable no-console */
 export async function register() {
   if (process.env.NEXT_RUNTIME === 'nodejs') {
-    const { runMigrations } = await import('./lib/client/run-migrations');
+    const { runMigrations } = await import(
+      './domain_core/infrastructure/run-migrations'
+    );
     try {
       await runMigrations();
       console.log('[Instrumentation] Migrations completed successfully');
